utils: SyscallReturnCode in utils/syscall.rs
Moved `SyscallReturnCode` struct and logic pertaining to this
struct from utils/src/lib.rs to utils/src/syscall.rs.
